APIService/celery/requirements.txt
modal
requests
celery
supabase
Pillow
openai